Refactor: 사용자 내대여정보 리팩토링
