+++ src/pages/History/LabRentalHistory/index.jsx
@@ -73,9 +73,9 @@ export default function LabRentalHistory() {
         <S.Title>내 대여 정보</S.Title>
         <S.RentalWrap>
           <h2>랩실 대여</h2>
-          <UserState isLab={true} />
+          <UserState />
           <h2>랩실 대여 이력</h2>
-          <UserHist isLab={true} />
+          <UserHist />
         </S.RentalWrap>
       </>
   );
